@@ -1,8 +1,10 @@
 module github.com/jupj/qvik
 
-go 1.13
+go 1.20
 
 require (
-	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4
-	golang.org/x/sys v0.0.0-20190412213103-97732733099d
+	golang.org/x/crypto v0.6.0
+	golang.org/x/sys v0.5.0
 )
+
+require golang.org/x/term v0.5.0 // indirect
