@@ -1,9 +1,12 @@
-// +build !windows
+//go:build !windows
 
 package main
 
-import "os"
-import "golang.org/x/crypto/ssh/terminal"
+import (
+	"os"
+
+	"golang.org/x/crypto/ssh/terminal"
+)
 
 // enableRawMode sets the console into raw mode and returns a restore function.
 // If err == nil, restore will restore the console to it's previous mode.
